requirements.txt
# Appium 相关
Appium-Python-Client==3.1.1
selenium==4.16.0

# iOS设备管理
tidevice==0.9.1

# 基础依赖
urllib3<2.0.0
requests==2.31.0
PyYAML==6.0.1
typing-extensions>=4.10.0

# 测试框架相关
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-rerunfailures==12.0
pytest-ordering==0.6
allure-pytest==2.13.2

# 日志相关
python-json-logger==2.0.7
colorlog==6.7.0

# Web UI 相关
streamlit==1.32.0

# 工具库
pillow==10.1.0  # 图片处理
decorator==5.1.1  # 装饰器工具
retry==0.9.2     # 重试机制

# 其他依赖
pandas>=2.0.0    # 数据处理
numpy>=1.24.0    # 数值计算
